feat!: move color definition into `theme`
The color definitions in main config are deprecated and removed. Now
these definitions doesn't have any affects.

Instead of this you should define your own color theme at `noti/themes/`
directory.
